Extract render column group callback
