+++ src/components/Spreadsheet/Spreadsheet-stories.module.sass
@@ -67,11 +67,16 @@
   justify-content: center
   cursor: pointer
 
-.group
+.group, .group div
   background-color: #f1f1f1
 
 .columnGroup
   composes: group
+
+.groupExpandCollapseButtonContainer
+  display: flex
+  justify-content: center
+  align-items: center
 
 .expandCollapseGroupButton
   composes: groupButton
