+++ src/components/Spreadsheet/index.js
@@ -2,4 +2,5 @@ export { default as useSpreadsheet } from './useSpreadsheet';
 export { default as SpreadsheetResizer } from './SpreadsheetResizer';
 export { default as SpreadsheetCell } from './SpreadsheetCell';
 export { default as SpreadsheetContaner } from './SpreadsheetContainer';
+export { default as renderColumnGroup } from './renderColumnGroup';
 export { default } from './Spreadsheet';
